+++ instrumentation/logback/logback-appender-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/logback/appender/v1_0/LogbackSingletons.java
@@ -28,8 +28,6 @@ public class LogbackSingletons {
         config.getBoolean("capture_marker_attribute/development", false);
     boolean captureTemplate = config.getBoolean("capture_template/development", false);
     boolean captureArguments = config.getBoolean("capture_arguments/development", false);
-    boolean captureLogstashStructuredArguments =
-        config.getBoolean("capture_logstash_structured_arguments/development", false);
     Predicate<String> mdcAttributes =
         SelectorConfig.resolveLegacyLiteral(config, "logback-appender", "mdc-attributes");
     Predicate<String> keyValuePairAttributes =
@@ -41,6 +39,12 @@ public class LogbackSingletons {
     Predicate<String> logstashMarkerAttributes =
         SelectorConfig.resolveLegacyBoolean(
             config, "logback-appender", "logstash-marker-attributes");
+    Predicate<String> logstashStructuredArgumentAttributes =
+        SelectorConfig.resolveLegacyBoolean(
+            config,
+            "logback-appender",
+            "logstash-structured-argument-attributes",
+            "logstash-structured-arguments");
 
     mapper =
         LoggingEventMapper.builder()
@@ -53,7 +57,7 @@ public class LogbackSingletons {
             .setCaptureTemplate(captureTemplate)
             .setCaptureArguments(captureArguments)
             .setLogstashMarkerAttributes(logstashMarkerAttributes)
-            .setCaptureLogstashStructuredArguments(captureLogstashStructuredArguments)
+            .setLogstashStructuredArgumentAttributes(logstashStructuredArgumentAttributes)
             .build();
   }